% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_trailing_zero.R
\name{drop_trailing_zero}
\alias{drop_trailing_zero}
\title{Drop Trailing Zero}
\usage{
drop_trailing_zero(x, align)
}
\arguments{
\item{x}{a numeric value or vector.}

\item{align}{alignment ('left' or 'right') of returned values.}
}
\value{
a character (single or vector depending on \code{x}).
}
\description{
Drops trailing zero(s) with no mathematical value.
}
\examples{
x <- 0.00200
drop_trailing_zero(x, aligh = "left")
drop_trailing_zero(x, aligh = "right")

x <- c(0.00200, 0.1, 1110.0250)
drop_trailing_zero(x, aligh = "left")
}
