% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medixrdemoa-package.R
\docType{package}
\name{medixrdemoa-package}
\alias{medixrdemoa}
\alias{medixrdemoa-package}
\title{medixrdemoa: Demo A}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The first demo for medixr package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mehdiaqdam/medixrdemoa}
  \item Report bugs at \url{https://github.com/mehdiaqdam/medixrdemoa/issues}
}

}
\author{
\strong{Maintainer}: Mehdi Aqdam \email{mehdimoslemiaqdam@gmail.com}

}
\keyword{internal}
